Hack for issue 165, don't assume samples have an equal weight
Summary:

Disclaimer - I am not an expert on Perl, perf or FlameGraph.

This is a hack for https://github.com/brendangregg/FlameGraph/issues/165
By hack I mean that it works for me but I didn't update all of the test files.
The problem is that stackcollapse assumes that all stack traces have an equal
weight but they don't based on based on current perf output. For example,
there can be lines like this where 90689, 32298287 and 1 are the weights
but stackcollapse assumes they all have weight 1, and by weight I mean the
equivalent of the number of times this stack trace happens. Thus, the relative
frequency of xpl_accept-3 is 1 / (90689 + 32289287 + 1).

xpl_accept-1 2777710 12891973.429160: 90689 cycles:
xpl_accept-2 2777710 12891975.431112: 32289287 cycles:
xpl_accept-3 2777710 12891975.431112: 1 cycles:

This is based on https://github.com/brendangregg/FlameGraph/pull/250 but
I am not sure that PR is correct, it is also ~20 diffs behind current FlameGraph.

Test Plan:

works for me

Reviewers:

Subscribers:

Tasks:

Tags: